Rewrite Provisioner in Typescript
